[!][add] add `hikami_core` crate and move modules to it
