+++ Cargo.toml
@@ -18,7 +18,18 @@ panic = 'abort'
 codegen-units = 1
 
 [workspace]
-members = [ "extension_manager" ]
+members = [ "extension_manager" , "hikami_core"]
+
+[workspace.dependencies]
+elf = { version = "0.7.2", default-features = false }
+fdt = "0.1.5"
+linked_list_allocator = "0.10.5"
+raki = "1.3.1"
+riscv = "0.11.1"
+rustsbi = "0.4.0"
+sbi-rt = "0.0.3"
+sbi-spec = { version = "0.0.8", features = [ "legacy" ] }
+spin = "0.9.8"
 
 [features]
 # debug log
@@ -29,14 +40,10 @@ identity_map = []
 enable_extension = [ "hikami_zicfiss" ]
 
 [dependencies]
+hikami_core = { path = "hikami_core" }
 hikami_zicfiss = { path = "../hikami_zicfiss", optional = true }
-extension_manager = { path = "extension_manager" }
-elf = { version = "0.7.2", default-features = false }
-fdt = "0.1.5"
-linked_list_allocator = "0.10.5"
-raki = "1.3.1"
-riscv = "0.11.1"
-rustsbi = "0.4.0"
-sbi-rt = "0.0.3"
-sbi-spec = { version = "0.0.8", features = [ "legacy" ] }
-spin = "0.9.8"
+
+elf = { workspace = true }
+fdt = { workspace = true }
+riscv = { workspace = true }
+linked_list_allocator = { workspace = true }
