+++ hikami_core/src/device/initrd.rs
