+++ hikami_core/src/device/virtio.rs
