+++ hikami_core/src/emulate_extension.rs
