+++ hikami_core/src/h_extension.rs
