+++ hikami_core/src/log.rs
