+++ hikami_core/src/memmap/constant.rs
